feat:[#4] Add the interests page layout
Created the Interests.jsx file in the pages folder and included the basic structure of the page. This includes three large boxes that represent the page overview, the ‘nature’ section, and the ‘baking’ section. Added a new class in the index.css file for these boxes called ‘interest-box’.

Added the interests page in my App.jsx file to allow for navigation between pages. Also included a link to the Interests page on the Home page so I can access it from my website directly.