feat:[#4] Add photos in Interests page
Created a ‘photos’ folder in the ‘public’ folder. Put four images in the photos folder.

Added all four photos in the Interests.jsx file. Added an ‘images’ class in the index.css file to allow for consistent formatting of the pictures. Made edits to the ‘interest-box’ class to help accommodate the formatting of the pictures.