feat:[#5] Add contact page layout
Created the Contact.jsx file in the pages folder and included the basic structure of the page. This includes the background and the introduction section.

Added the contact page in my App.jsx file to allow for navigation between pages. Also included a link to the Contact page on the Home.jsx page so I can access it from my website directly.