feat:[#5] Add hyperlinks to contact page
Added two new classes in index.css (a button class and a class for the boxes). Created 6 cards (boxes) to hold both information and a button linking to external sites. Used a grid class to help structure the boxes.

Included hyperlinks for each of the six buttons, linking to external sites. Included a title related to each link in each box.