feat:[#3] Add the skills page layout
Created the Skill.jsx file in the pages folder. Added the skills page in the App.jsx file by importing it then adding a new route for skills. Modified Home.jsx to link the skills page on the homepage.

Added the basic structure of the skills page in the Skills.jsx file. This includes the background and the introduction section.