feat:[#3] Add technology section in skills page
Added a section in Skills.jsx that shows the different programming languages/technologies I know. Includes a title and a grid of six boxes that each display a language I am familiar with.

Added two new classes in index.css. One for the main box used to hold this section, and one for the coloured boxes that show each programming language.