style:[#20] Modify home page links
Modified the links on the homepage that take users to other pages on my site, so it was more clear to the user what to click. Made the button smaller and a different colour.

In the Home.jsx file I changed each of the page cards (boxes) to include the button changes, and updated the corresponding text so it was clear what to press. I also edited the index.css file to remove the shadow effect from the larger card, that way the only effect is on the button.